rewrite firmware raster implementation
New serial protocol and buffering for raster data.

Raster data is no longer streamed via the serial interrupt while the
move is already in progress. Instead it's expected right after the
"line" command if the new "raster bytes" parameter was used.  The
move is then queued by the planner together with its raster buffer.

The raster buffer is limited to 60 bytes, which will take long enough
to execute to allow the CPU to schedule the next move and combine it
without slowing down while cruising at high speed (8000 mm/min should
work).

The block buffer size was decreased to make room for 4 raster buffers.
I think this will not become a problem, but if it does, we should
consider to enlarge the serial RX buffer instead, because it buffers
commands with fewer bytes per move and it stores raster data too.