serial protocol changes, updated raster backend
New "raster_line" command in driveboard.py to split long raster moves
into smaller ones.

And move the calculation of pulses_per_mm/pulse_length for non-raster
moves away from the firmware into the Python backend.  The firmware
no longer needs an "intensity" setting.

Rationale:

It would have been nice if the backend could just specify the power in
percent or the pulse duration in microseconds and not care about the
IRQ frequency of the driverboard. But for raster mode this is not
really possible, because the backend needs to send the pulse duration as
integers and needs to take into account the size of those integer steps.

If the backend needs to be aware of this anyway, it can do those
calculations for non-raster moves as well (and probably faster, too).